[BugFix] `tablet_writer_add_chunk` timeout leads data lost (#21185)
When the RPC `tablet_writer_add_chunk` is timeout, the write thread would exit and return the request status. 
But `status` in `PTabletWriterAddBatchResult` is unset, so a `Status::OK()` is returned. The corresponding transaction would commit successfully.
The routine load would update the consuming progress according to the status of the transaction.
For those transactions with timeout `tablet_writer_add_chunk` requests, the consuming progress would be updated wrongly and Kafka messages are probably lost.