[BugFix] Fix the bug where truncating a partition with a duplicate partition name will cause fe to crash. (#21809)
The partition names in StarRocks are case insensitive, but the truncation process is case sensitive, which means that `truncate table t partition (p1, P1)` will write 2 new partitions in EditLog, which will result in an NPE when replaying the truncation log in the process of getting the dataProperty by partitionId, because only one partition can be left in the table.

To fix this bug, use a case insensitive map to store the origin parititions.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit 7383dbcb07d74705d17f2609ec248a222fdc4213)