[Enhancement] Avoid using page cache when collecting statistics (#21801)
Currently, BE scans segment files using pagecache when collecting statistics, which is a waste of resources. In addition, when large-scale data is continuously imported, if the amount of new data to be imported reaches a certain proportion, full statistics collection will be triggered, and page cache will fill up when the segment file is read. As a result, the imported memory will easily exceed the limit. This PR adds a field to queryoptions that allows FE to determine whether a query uses page cache. For collecting statistics, we use this field to tell BE not to use page cache when executing the query.
----
Signed-off-by: Zaorang Yang <zaorangy@gmail.com>
(cherry picked from commit b0d0533f3e5d1af98b313ede9e9bc49e440835b3)

# Conflicts:
#	be/src/exec/pipeline/scan/olap_chunk_source.cpp
#	be/src/exec/vectorized/tablet_scanner.cpp
#	be/src/runtime/runtime_state.h
#	fe/fe-core/src/main/java/com/starrocks/qe/SessionVariable.java
#	fe/fe-core/src/main/java/com/starrocks/statistic/FullStatisticsCollectJob.java
#	gensrc/thrift/InternalService.thrift