[BugFix] fix orc pushdown filter failure
When ORC executes filter pushdown, assuming there are two predicates in the SQL query, when the function `filterOnPickStringDictionary` is executed, if predicate1 is present in `sdicts`, dictionary filtering can be performed and `dict_value_chunk->num_rows()` after filtering is greater than 0, then `_dict_filter_eval_cache` will not be cleared. However, if predicate2 is not present in `sdicts`, the dictionary filtering in `filterOnPickStringDictionary` will not be executed, and `eval_conjuncts_into_filter` will obtain `chunk_size` of 0, which will execute the next chunk retrieval operation. It can be seen that `_dict_filter_eval_cache` will not be properly cleared. When the next batch of stripe data is retrieved and `_orc_row_reader_filter->is_slot_evaluated(it.first)` returns true, it will skip verification even if it does not satisfy the predicate."

Signed-off-by: linenwei1 <linenwei1@jd.com>