[BugFix] Some_problem_when_backup_restore_using_local_path (21735)
Problem 1:
When fe upload the backup file, fe will check the authority information.
But if we use local path, authority is undefined and fe will get NPE problem.

Problem 2:
When we use local path for restore, BE will parse the wrong path for
the remote file and let the restore fail.

Solution:
1. check authority information.
2. fix the parse problem.

Signed-off-by: srlch <linzichao@starrocks.com>