[BugFix] Schema check error for restore (#21418) (#21621)
Problem:
When the restored table has existed, sr will check the schema between local table and snapshot table.
But the check has some bugs and will cause some problem:
1. If local table and snapshot table has same columns with different nullable attribute,
the check will success and it will cause crash because fe and be have different schema for nullable attribute.
2. Schema check include the value of schema hash, and this value will change when finishing schema change.
If a local table finish schema change and get the same schema compare with the remote table, it will still
fail in schema check because the schema hash is different, but it is no make sense.

Solution:
1. Check the BaseSchema between local and remote table.
2. skip the check for schema hash value if doing restore.

Signed-off-by: srlch <linzichao@starrocks.com>
Conflicts:
	fe/fe-core/src/main/java/com/starrocks/catalog/OlapTable.java