[BugFix] AUTO_INCREMENT insert/load error when table has nullable column (#21877)
Problem:
Currently, we use nullExprInAutoIncrement to represent that if user specify
null expression for auto increment column. But there is some bug for the
nullExprInAutoIncrement

Insert:
If the table with auto increment column and other nullable column, the nullExprInAutoIncrement
will be set even though the auto increment column is not set by null. This will cause insertion failure.

Load:
If the table with auto increment column and other nullable column. If user specify the NULL value for
auto increment column explictly, and miss the nullable column, the load will be successfully finished
but not cancel.This is unexpected behavior.

Solution:
1. Fix the nullExprInAutoIncrement problem.
2. refactor the code, throw exception when Insert/Update has nullExprInAutoIncrement == true

Signed-off-by: srlch <linzichao@starrocks.com>
(cherry picked from commit 295123083a881016a9167858d29ddf4d5c7843fb)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/sql/UpdatePlanner.java
#	fe/fe-core/src/main/java/com/starrocks/sql/analyzer/UpdateAnalyzer.java
#	fe/fe-core/src/main/java/com/starrocks/sql/ast/UpdateStmt.java