[Enhancement] limit max auto increment cache size (#21249) (#21251)
In current implementation, if the auto increment id counter is overflow, the insert/load must be fail.
And if we set the cache size too large, it will cause the overflow manually which is a problem here.
So the limitation must be set to prevent this situation. We can simply define the cache size as int instead of long

Signed-off-by: srlch <linzichao@starrocks.com>
(cherry picked from commit ba54aac8a399ab87849dd45faf405eb9d996bdcb)