[BugFix] Cast exception when compare two columns (#22134)
Problem:
In Column definition, we can use equals method to compare two column obj.
In this function, sr will compare strlen, precision and scale for ScalarType.
But when doing the comparison, the code will not check the column type whether
it is the ScalarType or not. If it is not the ScalarType, we will get the cast exception and fail.

Solution:
check the type is ScalarType or not.

Signed-off-by: srlch <linzichao@starrocks.com>