aarch64: use isb instead of yield
This patch is fixing a performance issue on arm64. The patch follows the same
pattern we have fixed in mysql and other software using spin loops.

A yield instruction in aarch64 is essentially a nop, and does not cause enough
delay to help backoff. isb is a barrier that, especially inside a loop, creates
a small delay without consuming ALU resources. Experiments shown that adding the
isb instruction improves stability and reduces result jitter. Adding more delay
to the UT_RELAX_CPU than a single isb reduces performance.

Signed-off-by: Sebastian Pop <spop@amazon.com>