[Enhancement] Fast orphan segment deletion of cloud native (#22519)
* [Enhancement] Fast orphan segment deletion of cloud native

During the process of tablet metadata GC, if it is found that a certain
metadata to be deleted is generated by compaction task, the input files
of that compaction is directly deleted. In this way, the space of
persistence storage and local disk cache can be released more immediately.

Signed-off-by: sduzh <zhuming9011@qq.com>
(cherry picked from commit 95930698df48c4ecd29f93cc01d57f4abac708fb)