[Enhancement] Optimize tablet scheduler for large pending tablets case (#21647)
Now, there are some bad cases in the logic of tablet scheduler when the number of tablets to be repaired is large: the priority of tablet will be downgraded if there is not enough resource(be path slot), and the tablet in low priority will never be scheduled if there are new tablets to repair continuously, because the new tablets will be put to the head of schedule queue. The tablet in low priority can be upgraded to high priority, but the time interval is very big, it's 30 mins.
To cover this bad case:
1. do not downgrade the priority of tablet if there is not enough resource, because the resource should be allocated to the selected tablet before first.
2. schedule all the tablet in queue one time, because the tablet on the head and tablet on the tail may use different be resource.
3. since the number of tablet to schedule in one batch is increased, we should optimize the granularity of db lock: use db read lock to check existence of meta, and use db write lock when add or drop replica.

Signed-off-by: gengjun-git <gengjun@starrocks.com>