[BugFix] Fix Report Tablet memory leak bug (#22772)
ReportHandler uses `reportQueue` to record the Report order, and uses `pendingTaskMap` to record the latest task, when processing the task, it fist pop one task (task1) from `reportQueue`, and get the latest task (task2) from `pendingTaskMap` of the same type and beId to task1. If the processing speed is slow, the task1 and task2 will be different task, which means there will be double number of TTabletInfo in the memory at the same time, it wastes too much memory.
To fix it, only record the reportType and beId in the `reportQueue`.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit 143c55344106e5908e33062ff54d44e5135a270d)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/leader/ReportHandler.java