[BugFix] Be maybe crash due to the incorrect overlap info (#23191)
We record the overlap info(including overlap kv num and overlap kv size) during L0 and L1, but we don't update the overlap info correctly when we try to delete the row whose primary key length is more than 64 Bytes. And this may cause BE to crash in the future merge compaction.