BugFix] remove useless hard link for file in clone and restore (#23588)
Problem:
Currently, be will reset rowset id and link the file using new rowset id
when execute clone and restore. But it cause a problem, it will link both
new_name and old_name to tablet dir but only the new name is vaild for the
rowset meta. If we want to remove the rowset , it can not delete the file
because we can not access the old_name and unlink it, so the number of the
file link is not zero in this case.

Solution:
unlink the old_name

Signed-off-by: srlch <linzichao@starrocks.com>