[Enhancement] Adapt tablet repair process to k8s+pvc recovery mode (#23654)
## Problem Summary:
Fixes
https://github.com/StarRocks/starrocks-kubernetes-operator/issues/49

For certain deployment, like k8s pods + pvc, the replica is not lost even the corresponding backend is detected as dead, because the replica data is persisted on a pvc which is backed by a remote storage service, such as AWS EBS. And later, k8s control place will schedule a new pod and attach the pvc to it which will restore the replica to a normal state immediately. But normally the tablet scheduler of Starrocks will start to schedule REPLICA_MISSING type tasks and create new replicas in a short time. After new pod scheduling is completed, tablet scheduler has to delete the redundant healthy replica which cause resource waste and may also affect the loading process.

We add a new configuration `tablet_sched_ignore_dead_backend` to adapt to this recovery mode, when setting it to true, StarRocks won't schedule any REPLICA_MISSING type task because of dead backends (but may still schedule REPLICA_MISSING type tasks because of other reasons, like manually setting a replica as bad, actively decommission a backend etc.) and leave the resource manager like k8s to handle the repair process.

Signed-off-by: Dejun Xia <xiadejun@starrocks.com>