[Enhancement] Eliminate list objects operations when obtaining tablet schemas
Create a schema file under the root path of the table. When you need to obtain
the schema of a tablet, first get the ID of the materialized index to which the
tablet belongs, and then read the schema file according to that ID to avoid the
list tablet metadata operation.

Signed-off-by: sduzh <zhuming9011@qq.com>