[Enhancement] Make orphan files deletion more reliable and faster for cloud native table (#23480)
Prefix the data file names with the transaction ID to make the orphan
files deletion more reliable and faster.

Use transaction IDs as logical file creation timestamps. During garbage
collection, if an unreferenced data file has a logical creation time
preceding the minimum ID of active transactions, it is deemed to not
belong to any in-progress transactions and can be safely purged.

Signed-off-by: sduzh <zhuming9011@qq.com>