[BugFix] The primary key table segment data maybe out of order after finish schema change (#23985)
The rowset of the primary key tablet after import may be segmented
overlapping until compaction is finished. But if we create an altered
job(add an index or change column type) that needs to rewrite the
segment data, we ignore the segment overlapping status and splice into a
new segment one by one according to the original order of the segments.
This may cause the new segment data out of order and the short key index
will be incorrect we may get an error query result using a short key
index.

---------

Signed-off-by: zhangqiang <qiangzh95@gmail.com>
(cherry picked from commit cc2929898987091683a681dd42e874d9bb61bb2d)

# Conflicts:
#	be/src/storage/tablet_updates.cpp
#	be/src/storage/tablet_updates.h