[Refactor] Move the STARROCKS_VERSION and STARROCKS_COMMIT_HASH to the build.sh (#24215)
Currently, the two variables are defined in the
`build-support/gen_build_version.py`. It's hard to debug and search. The
pull request move the logic to the build.sh. The definition logic of
STARROCKS_VERSION is as the following.
1. If the user set the STARROCKS_VERSION, use the user's definition.
2. If the current commit is a tag, use the tag name.
3. If no tag is available, use the branch name
4. If not a branch, use the commit-hash

(cherry picked from commit 5af915f39ba64aadf8708ed9c42a8fbc5f7971b6)

# Conflicts:
#	build-support/gen_build_version.py
#	build.sh