[Enhancement] Jsonpaths does not needed for avro union data (#23556)
## Problem Summary:
For avro data that does not have nested data, the user must specify the
jsonpaths when parsing union data. If you're loading a table with
hundreds or thousands of columns, you need to write jsonpath for each
column, which is tedious. This PR fixes this issue.

## What type of PR is this:
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto
backported to target branch
  - [x] 3.0
  - [ ] 2.5
  - [ ] 2.4
  - [ ] 2.3

---------

Signed-off-by: Zaorang Yang <zaorangy@gmail.com>