[Enhancement] Add broker load error in retrying (#21982)
## What type of PR is this：
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related
issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will
this bug be triggered and what measures have you taken to fix the bug?)
-->
The broker load job would retry for 3 times if the plan stage get any
problem, but the error message would not be shown until until the job is
canceled.
To make it simpler to find out the problem, This PR adds the error
message in the process of retrying to the result of `show load`.

![8lZWOjXia7](https://user-images.githubusercontent.com/16649269/233242625-3ad6a99b-ef04-4538-b4c5-59d142e02319.jpg)

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
 - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto
backported to target branch
  - [x] 3.0
  - [x] 2.5
  - [ ] 2.4
  - [ ] 2.3

Signed-off-by: ricky <rickif@qq.com>
(cherry picked from commit a6e6b89bb790ff6e27fed90f15ae0b3955029966)