[Enhancement] Add a lru insertion point option to block cache to support inserting the new cache item to a custom position of the lru list. (backport #23267) (#23632)
In this PR, we add an option to support cachelib lru insertion point,
which is a custom postion and not at the head of the list. New items
will be inserted at the position ipos(p) = size(LRU) * p from the tail,
where p ∈ [0,1] is a ratio parameter. For example, suppose the regular
LRU tail age be 600s, but you want to cache only items that were
accessed during the 200s after insertion. Inserting new items at
position ipos(1/3) from the tail may help you to achieve this. Not that
custom IP affects only items that're inserted into the cache the first
time; it doesn't affect items that on access are moved to the head of
the list.


![image](https://github.com/StarRocks/starrocks/assets/13727748/7166d799-0d69-44ac-9cf4-cb7637a53088)
Signed-off-by: GavinMar <yangguansuo@starrocks.com>