[Feature] Support CUME_DIST window function (#22775)
The CUME_DIST window function returns the cumulative distribution of a value within a group of values; that is, the percentage of partition values less than or equal to the value in the current row. This represents the number of rows preceding or peer with the current row in the window ordering of the window partition divided by the total number of rows in the window partition. Return values range from 0 to 1.

It can not been used with the windowing clause. And for the implementation, it uses `RANGE BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW`.

NOTE that, CUME_DIST need the boundary of partition to calculate its value. For these functions, we must wait util the partition finished. Therefore, it uses `_process_by_partition_for_unbounded_preceding_range_frame_materializing`.

The CUME_DIST window function is only supported by the pipeline engine. BE will report the NotSupported error, when non-pipeline engine executes CUME_DIST in BE.

---------

Signed-off-by: Moonm3n <saxonzhan@gmail.com>