[Enhancement] Add zone map for sort key column (#24593)
We support separate primary key and sort key for the PrimaryKey table so
far, but we don't create a ZoneMap index for the sort key column if the
sort key column is not the key column. So the query may need to read
more data and use predicates to filter rows. We should create a ZoneMap
index for sorting key columns too.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>