[BugFix] Fix checkpoint failed bug when writing cluster (#24612)
Fixes #https://starrocks.atlassian.net/browse/SR-18050

If there is a database named `starrocks` in the system, then upgrade to
version 3.0, the checkpoint will fail with the following log
```
2023-05-29 15:20:19,471 ERROR (leaderCheckpointer|368) [Checkpoint.replayAndGenerateGlobalStateMgrImage():226] Exception when generate new image file
java.lang.NullPointerException: null
        at com.starrocks.cluster.Cluster.write(Cluster.java:185) ~[starrocks-fe.jar:?]
        at com.starrocks.server.LocalMetastore.saveCluster(LocalMetastore.java:4650) ~[starrocks-fe.jar:?]
        at com.starrocks.server.GlobalStateMgr.saveImage(GlobalStateMgr.java:1702) ~[starrocks-fe.jar:?]
        at com.starrocks.server.GlobalStateMgr.saveImage(GlobalStateMgr.java:1667) ~[starrocks-fe.jar:?]
        at com.starrocks.leader.Checkpoint.replayAndGenerateGlobalStateMgrImage(Checkpoint.java:217) ~[starrocks-fe.jar:?]
        at com.starrocks.leader.Checkpoint.runAfterCatalogReady(Checkpoint.java:106) ~[starrocks-fe.jar:?]
        at com.starrocks.common.util.LeaderDaemon.runOneCycle(LeaderDaemon.java:73) ~[starrocks-fe.jar:?]
        at com.starrocks.common.util.Daemon.run(Daemon.java:115) ~[starrocks-fe.jar:?]
```
if the `dbNames` contains `starrocks`, the `dbNameToIDs` will not
contain the key of `starrocks` if `starrocks` is loaded from image,
because `dbNameToIDs` is not persist to image, so we should check the
existence of the key in `dbNameToIDs` before get it.
Since the cluster will be removed from 3.1 and the build-in db
`starrocks` is added in 3.0, we only need to fix it on 3.0.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit b55a118bb15d6d1adf61e74e761c77ce7c3896a1)