[Feature] Completely resolving the issue of query failures in Hive external tables due to inconsistent metadata caching (#24501)
Fixes #issue
If modifications are made to the partition files of a Hive table by
user, such as through "insert overwrite partition", the Frontend
couldn't be aware of these changes. As a result, queries may use the old
file information cached in the FE for execution. Be will throw exception
when accessing a non-exist remote files. So we need to resolve this
issue.
There are some alternative solutions:
- aws sqs/sns  // Cannot be aware of a complete transaction.
- glue notification // it can't work when user use hive metastore.
- hive metastore listener // user need to enable metastore events in the
hive metastore config and restart it. most of user can't do this
operation.
- Background Periodic Refresh // The refresh interval is long, and users
still have a high probability of encoutering query errors. The refresh
interval is short, background tasks put a lot of pressure on hive
metastore and namenode.

I think the most out-of-the-box solution is to retry the query. When the
file can't be found, refresh the metadata cache and regenerate the scan
range.

Signed-off-by: stephen <stephen5217@163.com>
(cherry picked from commit c2e911b552d332603f661eefb1985cf6589761dd)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/planner/HudiScanNode.java
#	fe/fe-core/src/main/java/com/starrocks/qe/Coordinator.java
#	fe/fe-core/src/main/java/com/starrocks/qe/StmtExecutor.java