[Enhancement] Make l0 snapshot size configurable (#24748)
Fixes #issue

Make l0 snapshot size configurable.

In a scenario with a large number of tablets, such as a scenario where
there are many wide tables and many tablets, one tablet is relatively
large, but the PrimaryIndex is around 10M, which cannot trigger the
flush l1 condition, resulting in a large memory usage.