[Feature] split scan or predicate to union (#24840)
### Main work
For sql like `select * from tbl where col1 > 100 and (col2 = 1 or col3
=2)`, the `(col2 = 1 or col3 =2)` predicate can actually filter lots of
rows, but our index doesn't support this predicate. We can transform it
to:
```
select * from tbl where  col1 > 100 and col2 = 1
union all
select * from tbl where col1 > 100 and col3 = 2 and (col2 !=1 or col2 is null);
```
In the second child, `col3 = 2` helps filter most rows and the other
preidicate can based on its result do the rest filter job.

Why not make modifications on the BE? There are several reasons:
1. Modifying existing indexes and late materialization to support OR
expressions is a complex work.
2. In clients' scenarios, predicates can be complex, and the selectivity
and statistical information are closely related. It is more efficient
for the FE to filter out predicates that meet the "good case" criteria
and simplify them into a standardized form before handing them over to
the BE for processing.

### split strategy
1. Check if the scan data size is greater than
`SCAN_OR_TO_UNION_THRESHOLD` with a default value of 1000000 rows * 50
bytes.
2. Split the conjunctions and determine if there already exists
predicates with better selectivity. Only predicates that use a single
column and have a selectRatio < 0.035 are considered to have good
filtering selectivity.
3. If there are no predicates with good selectivity, further split the
OR predicates. Check each element in the OR predicates to see if it
qualifies as a predicate with good selectivity. If all elements qualify
and the size of elements is samller than `SCAN_OR_TO_UNION_LIMIT`, split
the OR predicate.

BTW, add a speicial process in clause with lots of values, we will split
the values more than 1024 into some sets with most 1024 values helps use
the index.