[Enhancement] Support rewrite new sync mv by MaterializedView rule (#24764)
Fixes #22401

Sync MV is currently rewritten by the MaterializedViewRule, where the
rule retrieves the sync MV's columns based on a specific pattern.

This pull request (PR) introduces the usage of the new
BaseMaterializedViewRewriteRule.java for rewriting the sync MV in common
cases, similar to how it is used for async MV.

The main objective is to convert the existing sync MV into a new
MaterializedView format so that it can be utilized by the new rule,
BaseMaterializedViewRewriteRule. This conversion allows for the
application of the updated rule to the sync MV.

After this pr, sync mv can support more complex expressions later.

Signed-off-by: Kevin Li <ming.moriarty@gmail.com>