[BugFix] reset index id in MaterializedIndexMeta when restore (#25154)
Problem:
Restore will reset index id in MaterializedIndex, but currently does not
reset the corresponding index id in MaterializedIndexMeta. This will cause
the problem that when the table has some rollup index, and we backup and
restore it. It will get NPE when we query it. Because, we want to rewrite
the query using materialized view but we have no correct id information in
MaterializedIndexMeta.

Solution:
Reset index id in MaterializedIndexMeta when doing restore

Signed-off-by: srlch <linzichao@starrocks.com>