[BugFix] stream load failed when table with generate column (#25208)
Problem:
When use streamload without specifying the column name, generated column
will be treated as a normal column which can be insert by streamload.
This is an unexpected behavior.

Solution:
If user does not specify the column name in streamload, just load the data
by the columns without generated column.

Signed-off-by: srlch <linzichao@starrocks.com>