[Feature] Support add multiple generated column and reorder when adding normal column (#24885)
This pr solve two problem:

Problem 1:
Currently, if we add a normal column into a table with generated column. The normal column will
be added automatically before the generated column. We will keep the order of the columns that
generated columns must be at the last. But it is not work if we add multiple normal column at
once. The pr fix the problem.

Problem 2:
Support add multiple generated column using at once as following:
ALTER TABLE t ADD COLUMN (newcol1 INT AS id, newcol2 as id * 10)

Signed-off-by: srlch <linzichao@starrocks.com>