[BugFix] Primary key table data separated by primary key and sort key may be disordered after compaction (#25188)
If we create a primary key table that separates the primary key and the
sort key, we should order segment data by the sort key but not the
primary key. But when compaction is finished, the segment data may be
out of order because in the heap merge of segment data, the
`HeapMergeIterator` use the primary key as the basis for sorting data.