[BugFix] stream load failed when table with auto increment column(#25301)
Problem:
When use streamload without specifying the column name, auto increment
column
will be treated as a normal column which can be insert by streamload.
This is an unexpected behavior.

Solution:
If user does not specify the column name in streamload, just load the
data by the columns without auto increment column.

Signed-off-by: srlch <linzichao@starrocks.com>