[Enhancement] Limit the block_size for block cache. (#25228)
The cachelib doesn't support a item (key+valueu+attribute) larger than 4
MB without chain. So, we check and limit the block_size configured by
users to avoid unexpected errors. For starcache, in theory we doesn't
have a hard limitation for block size, but a very large block_size may
cause heavy read amplification. So, we also limit it to 2 MB as an
empirical value.

Signed-off-by: GavinMar <yangguansuo@starrocks.com>