[Enhancement] Optimize and modify the refresh schedule (#25399)
Fixes #issue

The behavior of the previous alter mv refresh was achieved by dropping a
task and then recreating a task. The problem with this implementation is
that in a concurrent scenario, when the user uses alter mv refresh, all
pending task_runs will be cleared when the drop task is triggered,
thereby resulting in a partial refresh.
In order to solve this problem, we do not use the method of drop task,
but reserve an alter task interface through the task, and directly
modify the scheduling of the task, so that the task_runs in the queue
will not be cleared when the modification is refreshed, so that the task
can be executed normally.

Signed-off-by: Astralidea <astralidea@163.com>
(cherry picked from commit e71416b69dd63a4d652941923ecb24d68b9eb856)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/alter/Alter.java
#	fe/fe-core/src/main/java/com/starrocks/persist/EditLog.java