[BugFix] Fix listTableSatus() NPE when base table is dropped (#25472)
Fixes SR-18423

When the base table of a view is dropped, NPE maybe throwed
when calling FrontendServiceImpl.listTableStatus() to list views
in a database because of privilege checking. In this case, we
should just ignore the privilege checking for the dropped table.

Signed-off-by: Dejun Xia <xiadejun@starrocks.com>
(cherry picked from commit d894c5ce7cbed9f627fe1151d51ab633732670c2)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/authentication/AuthenticationManager.java
#	fe/fe-core/src/main/java/com/starrocks/service/FrontendServiceImpl.java
#	fe/fe-core/src/main/java/com/starrocks/service/InformationSchemaDataSource.java