[Refactor] refactor spill trigger strategy for hash join (#25191)
We try to reserve a large amount of memory (1G per operator) when
restoring the partition hash table. In this PR, we modify the memory
table size to memory_table_num * memory_table_bytes_size. If each
partition is currently small, but the overall memory table exceeds the
memory_table_bytes_size, we will flush half of the memory table to
ensure that the memory table is smaller than the expected size.

Signed-off-by: stdpain <drfeng08@gmail.com>