[BugFix] Fix drop FE failed when the target FE is using FQDN bug the leader node is not (#25575)
Fixes SR-18292
Currently, the leader will resolve the host to IP when dropping FE if
the leader is not using FQDN mode. But if the target node is using FQDN,
the operation will fail because the host stored in meta is FQDN.
To fix this bug, only resolve host to IP when adding node (FE or BE),
and remain the host unchanged when dropping node.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit 82f880c6e6a7ea96b515ac8d20be45bf82376482)