[BugFix] The short key index maybe not order by sort key columns (#25591)
This bug is introduced by this
pr(https://github.com/StarRocks/starrocks/pull/25286), we fix the bug
which short key index and segment data may not be consistent after
vertical compaction in the pr. But the fix ignores the following
scenario:
if we create a table which order by column `c2` and c1 and `c2` is after
`c1`,
```
CREATE TABLE pk1sk2_tbl (
    k1 int,
    c1 varchar(20),
    c2 varchar(20)
PRIMARY KEY(k1)
DISTRIBUTED BY HASH(c1) BUCKETS 3
ORDER BY(c2,c1);
```
The data of column `c2` should be the prefix of short key index, but the
following code in `segment_writer.cpp`:
```
    for (uint32_t i = 0; i < _column_indexes.size(); i++) {
        uint32_t column_index = _column_indexes[i];
        ....
        if (column.is_sort_key()) {
            _sort_column_indexes.emplace_back(i);
        }
    }
```
When we import data or do horizontal compaction, segment_writer's init
schema is tablet schema, and the `_column_indexes` is ordered by table
column. At this time, column `c1` and column `c2` are both sort key
columns and we check column `c1` first, so the `_sort_column_indexes`
keeps the index of `c1` first. And in the following short key index
encoding process, the data of `c1` will be written before `c2` and be
the prefix of the short key index.

The solution is to check `_sort_column_indexes` and the elements of
`_sort_column_indexes` should be ordered by the sort key column index.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>