[Enhancement][Cherry-Pick][Branch-3.0] Support delete sql for primary key table with sort key (#22799) (#25644)
Primary key table with sort key does not support partial update so far,
so we need to provide all columns data when we doing data import. But
the delete sql only provide primary key columns and the delete request
will be reject because the request is consider as partial update
request. This pr will support the delete request for primary key table
with sort key.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>