[BugFix] Ensure workgroup is always set in BE (#25847)
The change introduced in #24421 removes the non-resource-group executor
and always utilizes the resource-group executor.

However, if the FE passes a false value for
`query_options.enable_resource_group` to the BE, the BE will fail to set
a workgroup for this fragment instance, leading to a crash. This
situation may arise when the FE version is older than #24421, but the BE
version is newer than #24421.

Signed-off-by: ZiheLiu <ziheliu1024@gmail.com>