[Refactor] Make casts of enum class to int explicit in fmt::format(). (#25806)
The old fmt-7.0.3 being used in Starrocks converted enum class to int
implicitly.
But it is not allowed anymore in the recent fmt library and will cause a
compile error.
So this explicit cast is a prerequisite for fmt library upgrade which
I'm planning to request in a separate PR.

Signed-off-by: jkim650 <117620956+jkim650@users.noreply.github.com>
(cherry picked from commit b092a91bbdd3042afa69113be92a5fe33090f4ed)