[BugFix] Fix some bug of executing alter operation for tables separating primary key and sort key (#25774)
This pr fix two bugs:
1. Adding/dropping columns for tables with separate primary and sort
keys could cause short index errors
2. Modifying the position of a non-sorted table could change the sorted
key column

The reason of the bugs is we do not check if the origin tablet has sort
key or not and use primary key to calculate short key column num. But
the sort key column index still be assign to new created tablet.