[Feature] Support runtime profile (#25131)
Sometimes, when a query is expected to take a considerable amount of
time to execute, such as 10 minutes, we are required to wait until its
completion before obtaining any information about the execution process.

Therefore, it is crucial to have a runtime profile available while the
query is still in progress.

We continue to use the existing session variable `enable_profile` to
enable/disable this feature. Besides, we introduce another session
varaible `runtime_profile_report_interval`, which defines the time
interval between two consecutive profile reports, with a default value
of 10.

The implementation is quite straightforward:
* Each fragment instance report its execution state and profile with the
defined interval `runtime_profile_report_interval`
* Fe receive the state report, and update the final query content with
the defined interval `runtime_profile_report_interval`

- [ ] BugFix
- [x] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

- [x] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [x] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4

---------

Signed-off-by: liuyehcf <1559500551@qq.com>