[Feature] Support list partition for materialized view. (#25783)
Fixes #issue

This PR is mainly to support Hive's Varchar type partition.
Since mv can be understood as an internal table. The internal table
currently supports Range better, but does not support List well.
However, the varchar type can only be realized through List
partitioning. So the essence of the problem here is that materialized
views support the creation of List partitions. And it supports refresh
of List partition. This also includes insert overwrite, which must also
support the refresh of the List partition.

The imported automatic partition table is not used here, because this
kind of table has limited support for many parameters, so it still
belongs to the mode of calculating the partition in advance according to
the plan and then refreshing it.

And there is still a lot of work to do here, which will be done in the
follow-up PR. For example to support other types. The improvement of
list partition. Support for internal list tables, etc.

Signed-off-by: Astralidea <astralidea@163.com>