[BugFix] Disable creating generated columns in lake and agg table (#26084) (#26085)
Problem:
1. Currently, generated column does not support lake table, but there is
no restriction for creating it in lake table which is a bug.
2. agg table can be created generated column actually, but it is
difficult to understand the semantic and it is no clearly scenes to use
it. So we should disable it too.

Signed-off-by: srlch <linzichao@starrocks.com>
(cherry picked from commit 7aa4b73c459813bd2619b6e51c1f81f7e8c41d1c)