[UT] Add UT to verify schemascannode execute on correct BE (#26573)
This PR is used to verify bug described in #26437, the reason for the
issue is the plan fragment was not executed on the expected BE.
Unfortunately, the issue was not reproducible on main and branch 2.5.
But still, the UT is helpful.

Signed-off-by: Binglin Chang <decstery@gmail.com>
(cherry picked from commit 08088084d352c3ddb35dcc5b4d2bfc6957687429)