[Enhancement] Add LastReportVersion to Replica to help detect data lose on backend (#26518)
Add lastReportVersion to Replica, and this version is updated only when
BE reports tablet, if the version on BE is less than lastReportVersion,
that means data is lost on BE, because this version can only be
increased monotonically.

Why not just check based on Replica.version?
There may be data load when processing the tablet report, in this case,
the version reported from BE is less than Replica.version too.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit 5d64387d2abc975c9842f00f399c110645b7bbc3)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/catalog/TabletInvertedIndex.java
#	fe/fe-core/src/main/java/com/starrocks/leader/ReportHandler.java