[BugFix] remove line separator correctly when log stmt (#26596)
The client may send query using a connector in windows platform and the
line separtator is `\r\n`. When we record this query in audit log, we
just replace the `\n` but not replace `\r`. In Java, \r is an escape
character that represents a carriage return. When you use `\r` within a
string, it tells the output device (such as a terminal or console) to
move the cursor to the beginning of the current line.
Like this
```
System.out.println("select \r from");
# only outoput
 from
```

Signed-off-by: packy <wangchao@starrocks.com>
(cherry picked from commit 6d1d7f728ca1071c3a083a2934fea18f4cba8a3c)