[Feature] support http query api (#25760)
This is a new implementation of http query api on the basis of
https://github.com/StarRocks/starrocks/pull/4464.
Instead of transforming result to json in FE, this pr generate result in
json format in BE, FE just send json to client, which can reduce Fe's
burden. This pr also register connectContext into connectScheduler, so
query timeout and wait timeout can be used for http sql. Right now
select, show ,explain, kill statements are supported, under the current
code framework, support other statement is very easy.

---------

Signed-off-by: mchades <793098424@qq.com>
Signed-off-by: before-Sunrise <unclejyj@gmail.com>
Co-authored-by: mchades <793098424@qq.com>