[Feature] Pruning tables in cardinality-preserving joins (#24843)
For cardinality-preserving joins:
1. A inner join B on A.fk = B.pk:  each row in A matches B exactly once;
2. A left join B on A.fk = B.pk: each row in A matches B at most once;

So the number of this join's output rows is equivalent to A's. if B's
columns is unused or it can be substituted by A's column that equivalent
to B's column, then B can be pruned from this join.
---------

Signed-off-by: satanson <ranpanf@gmail.com>
(cherry picked from commit 005b387a56750e1fad709ac9ca014fe40fbc3195)