[BugFix] crash when apply in table with AUTO_INCREMENT column (#27176) (#27199)
problem:
suppose we have a table created by:

```
CREATE TABLE t (id BIGINT NOT NULL,  name BIGINT NOT NULL,  v1 string, v2 string null, v3 BIGINT AUTO_INCREMENT) PRIMARY KEY (id, name) DISTRIBUTED BY HASH(id) BUCKETS 1 PROPERTIES(replication_num = 1, replicated_storage=true);
```

and then insert into a new record:
```
INSERT INTO t VALUES (1,2,3,4, DEFAULT);
```

and we have a CSV file with: 1,2,3 and 2,3,4

we want to partial update the v2 and v3, and we will get a crash when we
apply.

The reason is that, in the apply phase, we need to get the column id in
auto increment to do some work, but the id is incorrect in this case.

Fixes #27176

Signed-off-by: srlch <linzichao@starrocks.com>
Signed-off-by: The85thLetter <m.brian0606@gmail.com>