[BugFix] filter invalid plan generated by reorder algorithm (#27472)
Fixes #14257
For sql like:
```
select
  ref_0.N_NAME as c0,
  case when ref_3.C_NATIONKEY > ref_3.C_ACCTBAL then ref_1.O_ORDERPRIORITY else ref_1.O_SHIPPRIORITY end as c1
from
  nation as ref_0 inner join orders as ref_1
            on (ref_0.N_REGIONKEY = ref_1.O_CUSTKEY )
          left join t2 as ref_2
          on (ref_0.N_NATIONKEY = ref_2.v8 )
        inner join customer as ref_3
        on (ref_1.O_ORDERPRIORITY = ref_3.C_NAME )
      inner join supplier as ref_4
      on (ref_1.O_CLERK = ref_4.S_PHONE )
    left join supplier as ref_5
    on (ref_2.v9 = ref_5.S_SUPPKEY )
where cast(coalesce(ref_3.C_MKTSEGMENT, ref_0.N_NAME) as VARCHAR) = ref_5.S_ADDRESS
```
The logical plan with a on predicate `46: coalesce = 38: S_ADDRESS `
generated by two tables `ref_3` and `ref_0`. So when reorder, if we need
build a join with on predicate `46: coalesce = 38: S_ADDRESS`, the
`ref_3` and `ref_0` should be in oneside of the join. But our reorder
algorithm may generate a plan which the outermost join's left child
contains `ref_3` and its right child contains `ref_0` which failes to
yield a valid plan.
This pr add a check step to filter these planes. For `DP algorithm`, it
enumerate all posibility, so it can yield at least one valid plan. For
other algorithm, it may yield empty result.

Anohter changes:
Forcibly update rowOutputInfo when there is a projection, to avoid
invalidation of rowOutputInfo because only the projection is updated in
the transformation rule.

---------

Signed-off-by: packy92 <wangchao@starrocks.com>
Signed-off-by: packy <wangchao@starrocks.com>
Signed-off-by: The85thLetter <m.brian0606@gmail.com>