[BugFix] crash when apply in table with AUTO_INCREMENT column (#27176)
problem:
suppose we have a table created by:

and then insert into a new record:

and we have a CSV file with: 1,2,3 and 2,3,4

we want to partial update the v2 and v3, and we will get a crash when we
apply.

The reason is that, in the apply phase, we need to get the column id in
auto increment to do some work, but the id is incorrect in this case.

Signed-off-by: srlch <linzichao@starrocks.com>